Split task and add regression demo  (#130)
* split task

* add regression demo
