+++ paddlehub/finetune/task/basic_task.py
@@ -38,12 +38,6 @@ from paddlehub.finetune.evaluate import chunk_eval, calculate_f1, calculate_f1_n
 from paddlehub.finetune.config import RunConfig
 from scipy.stats import spearmanr
 
-__all__ = [
-    "ClassifierTask", "ImageClassifierTask", "TextClassifierTask",
-    "SequenceLabelTask", "MultiLabelClassifierTask", "ReadingComprehensionTask",
-    "RegressionTask"
-]
-
 
 class RunState(object):
     def __init__(self, length):
@@ -593,6 +587,7 @@ class BasicTask(object):
             run_states = []
             if self.current_epoch <= self.config.num_epoch:
                 while self.current_epoch <= self.config.num_epoch:
+                    self.config.strategy.step()
                     run_states = self._run(do_eval=do_eval)
                     self.env.current_epoch += 1
 
@@ -736,610 +731,3 @@ class BasicTask(object):
                 break
 
         return global_run_states
-
-
-class ClassifierTask(BasicTask):
-    def __init__(self,
-                 feature,
-                 num_classes,
-                 feed_list,
-                 data_reader,
-                 startup_program=None,
-                 config=None,
-                 hidden_units=None,
-                 metrics_choices="default"):
-        if metrics_choices == "default":
-            metrics_choices = ["acc"]
-
-        main_program = feature.block.program
-        super(ClassifierTask, self).__init__(
-            data_reader=data_reader,
-            main_program=main_program,
-            feed_list=feed_list,
-            startup_program=startup_program,
-            config=config,
-            metrics_choices=metrics_choices)
-
-        self.feature = feature
-        self.num_classes = num_classes
-        self.hidden_units = hidden_units
-
-    def _build_net(self):
-        cls_feats = self.feature
-        if self.hidden_units is not None:
-            for n_hidden in self.hidden_units:
-                cls_feats = fluid.layers.fc(
-                    input=cls_feats, size=n_hidden, act="relu")
-
-        logits = fluid.layers.fc(
-            input=cls_feats,
-            size=self.num_classes,
-            param_attr=fluid.ParamAttr(
-                name="cls_out_w",
-                initializer=fluid.initializer.TruncatedNormal(scale=0.02)),
-            bias_attr=fluid.ParamAttr(
-                name="cls_out_b", initializer=fluid.initializer.Constant(0.)),
-            act="softmax")
-
-        self.ret_infers = fluid.layers.reshape(
-            x=fluid.layers.argmax(logits, axis=1), shape=[-1, 1])
-
-        return [logits]
-
-    def _add_label(self):
-        return [fluid.layers.data(name="label", dtype="int64", shape=[1])]
-
-    def _add_loss(self):
-        ce_loss = fluid.layers.cross_entropy(
-            input=self.outputs[0], label=self.labels[0])
-        return fluid.layers.mean(x=ce_loss)
-
-    def _add_metrics(self):
-        acc = fluid.layers.accuracy(input=self.outputs[0], label=self.labels[0])
-        return [acc]
-
-    @property
-    def fetch_list(self):
-        if self.is_train_phase or self.is_test_phase:
-            return [self.labels[0].name, self.ret_infers.name
-                    ] + [metric.name
-                         for metric in self.metrics] + [self.loss.name]
-        return [output.name for output in self.outputs]
-
-    def _calculate_metrics(self, run_states):
-        loss_sum = acc_sum = run_examples = 0
-        run_step = run_time_used = 0
-        all_labels = np.array([])
-        all_infers = np.array([])
-
-        for run_state in run_states:
-            run_examples += run_state.run_examples
-            run_step += run_state.run_step
-            loss_sum += np.mean(
-                run_state.run_results[-1]) * run_state.run_examples
-            acc_sum += np.mean(
-                run_state.run_results[2]) * run_state.run_examples
-            np_labels = run_state.run_results[0]
-            np_infers = run_state.run_results[1]
-            all_labels = np.hstack((all_labels, np_labels.reshape([-1])))
-            all_infers = np.hstack((all_infers, np_infers.reshape([-1])))
-
-        run_time_used = time.time() - run_states[0].run_time_begin
-        avg_loss = loss_sum / run_examples
-        run_speed = run_step / run_time_used
-
-        # The first key will be used as main metrics to update the best model
-        scores = OrderedDict()
-
-        for metric in self.metrics_choices:
-            if metric == "acc":
-                avg_acc = acc_sum / run_examples
-                scores["acc"] = avg_acc
-            elif metric == "f1":
-                f1 = calculate_f1_np(all_infers, all_labels)
-                scores["f1"] = f1
-            elif metric == "matthews":
-                matthews = matthews_corrcoef(all_infers, all_labels)
-                scores["matthews"] = matthews
-            else:
-                raise ValueError("Not Support Metric: \"%s\"" % metric)
-
-        return scores, avg_loss, run_speed
-
-
-ImageClassifierTask = ClassifierTask
-
-
-class TextClassifierTask(ClassifierTask):
-    def __init__(self,
-                 feature,
-                 num_classes,
-                 feed_list,
-                 data_reader,
-                 startup_program=None,
-                 config=None,
-                 hidden_units=None,
-                 metrics_choices="default"):
-
-        if metrics_choices == "default":
-            metrics_choices = ["acc"]
-        super(TextClassifierTask, self).__init__(
-            data_reader=data_reader,
-            feature=feature,
-            num_classes=num_classes,
-            feed_list=feed_list,
-            startup_program=startup_program,
-            config=config,
-            hidden_units=hidden_units,
-            metrics_choices=metrics_choices)
-
-    def _build_net(self):
-        cls_feats = fluid.layers.dropout(
-            x=self.feature,
-            dropout_prob=0.1,
-            dropout_implementation="upscale_in_train")
-
-        if self.hidden_units is not None:
-            for n_hidden in self.hidden_units:
-                cls_feats = fluid.layers.fc(
-                    input=cls_feats, size=n_hidden, act="relu")
-
-        logits = fluid.layers.fc(
-            input=cls_feats,
-            size=self.num_classes,
-            param_attr=fluid.ParamAttr(
-                name="cls_out_w",
-                initializer=fluid.initializer.TruncatedNormal(scale=0.02)),
-            bias_attr=fluid.ParamAttr(
-                name="cls_out_b", initializer=fluid.initializer.Constant(0.)),
-            act="softmax")
-
-        self.ret_infers = fluid.layers.reshape(
-            x=fluid.layers.argmax(logits, axis=1), shape=[-1, 1])
-
-        return [logits]
-
-
-class SequenceLabelTask(BasicTask):
-    def __init__(self,
-                 feature,
-                 max_seq_len,
-                 num_classes,
-                 feed_list,
-                 data_reader,
-                 startup_program=None,
-                 config=None,
-                 metrics_choices="default"):
-        if metrics_choices == "default":
-            metrics_choices = ["f1", "precision", "recall"]
-
-        main_program = feature.block.program
-        super(SequenceLabelTask, self).__init__(
-            data_reader=data_reader,
-            main_program=main_program,
-            feed_list=feed_list,
-            startup_program=startup_program,
-            config=config,
-            metrics_choices=metrics_choices)
-        self.feature = feature
-        self.max_seq_len = max_seq_len
-        self.num_classes = num_classes
-
-    def _build_net(self):
-        self.logits = fluid.layers.fc(
-            input=self.feature,
-            size=self.num_classes,
-            num_flatten_dims=2,
-            param_attr=fluid.ParamAttr(
-                name="cls_seq_label_out_w",
-                initializer=fluid.initializer.TruncatedNormal(scale=0.02)),
-            bias_attr=fluid.ParamAttr(
-                name="cls_seq_label_out_b",
-                initializer=fluid.initializer.Constant(0.)))
-
-        self.ret_infers = fluid.layers.reshape(
-            x=fluid.layers.argmax(self.logits, axis=2), shape=[-1, 1])
-        ret_infers = fluid.layers.assign(self.ret_infers)
-
-        self.seq_len = fluid.layers.data(
-            name="seq_len", shape=[1], dtype='int64')
-        seq_len = fluid.layers.assign(self.seq_len)
-
-        logits = self.logits
-        logits = fluid.layers.flatten(logits, axis=2)
-        logits = fluid.layers.softmax(logits)
-        self.num_labels = logits.shape[1]
-        return [logits]
-
-    def _add_label(self):
-        label = fluid.layers.data(
-            name="label", shape=[self.max_seq_len, 1], dtype='int64')
-        return [label]
-
-    def _add_loss(self):
-        labels = fluid.layers.flatten(self.labels[0], axis=2)
-        ce_loss = fluid.layers.cross_entropy(
-            input=self.outputs[0], label=labels)
-        loss = fluid.layers.mean(x=ce_loss)
-        return loss
-
-    def _add_metrics(self):
-        self.ret_labels = fluid.layers.reshape(x=self.labels[0], shape=[-1, 1])
-        return [self.ret_labels, self.ret_infers, self.seq_len]
-
-    def _calculate_metrics(self, run_states):
-        total_infer = total_label = total_correct = loss_sum = 0
-        run_step = run_time_used = run_examples = 0
-        for run_state in run_states:
-            loss_sum += np.mean(run_state.run_results[-1])
-            np_labels = run_state.run_results[0]
-            np_infers = run_state.run_results[1]
-            np_lens = run_state.run_results[2]
-            label_num, infer_num, correct_num = chunk_eval(
-                np_labels, np_infers, np_lens, self.num_labels,
-                self.device_count)
-            total_infer += infer_num
-            total_label += label_num
-            total_correct += correct_num
-            run_examples += run_state.run_examples
-            run_step += run_state.run_step
-
-        run_time_used = time.time() - run_states[0].run_time_begin
-        run_speed = run_step / run_time_used
-        avg_loss = loss_sum / run_examples
-
-        precision, recall, f1 = calculate_f1(total_label, total_infer,
-                                             total_correct)
-        # The first key will be used as main metrics to update the best model
-        scores = OrderedDict()
-
-        for metric in self.metrics_choices:
-            if metric == "precision":
-                scores["precision"] = precision
-            elif metric == "recall":
-                scores["recall"] = recall
-            elif metric == "f1":
-                scores["f1"] = f1
-            else:
-                raise ValueError("Not Support Metric: \"%s\"" % metric)
-
-        return scores, avg_loss, run_speed
-
-    @property
-    def feed_list(self):
-        feed_list = [varname for varname in self._base_feed_list]
-        if self.is_train_phase or self.is_test_phase:
-            feed_list += [self.labels[0].name, self.seq_len.name]
-        else:
-            feed_list += [self.seq_len.name]
-        return feed_list
-
-    @property
-    def fetch_list(self):
-        if self.is_train_phase or self.is_test_phase:
-            return [metric.name for metric in self.metrics] + [self.loss.name]
-        elif self.is_predict_phase:
-            return [self.ret_infers.name] + [self.seq_len.name]
-        return [output.name for output in self.outputs]
-
-
-class MultiLabelClassifierTask(ClassifierTask):
-    def __init__(self,
-                 feature,
-                 num_classes,
-                 feed_list,
-                 data_reader,
-                 startup_program=None,
-                 config=None,
-                 hidden_units=None,
-                 metrics_choices="default"):
-        if metrics_choices == "default":
-            metrics_choices = ["auc"]
-
-        main_program = feature.block.program
-        super(MultiLabelClassifierTask, self).__init__(
-            data_reader=data_reader,
-            feature=feature,
-            num_classes=num_classes,
-            feed_list=feed_list,
-            startup_program=startup_program,
-            config=config,
-            hidden_units=hidden_units,
-            metrics_choices=metrics_choices)
-        self.class_name = list(data_reader.label_map.keys())
-
-    def _build_net(self):
-        cls_feats = fluid.layers.dropout(
-            x=self.feature,
-            dropout_prob=0.1,
-            dropout_implementation="upscale_in_train")
-
-        if self.hidden_units is not None:
-            for n_hidden in self.hidden_units:
-                cls_feats = fluid.layers.fc(
-                    input=cls_feats, size=n_hidden, act="relu")
-
-        probs = []
-        for i in range(self.num_classes):
-            probs.append(
-                fluid.layers.fc(
-                    input=cls_feats,
-                    size=2,
-                    param_attr=fluid.ParamAttr(
-                        name="cls_out_w_%d" % i,
-                        initializer=fluid.initializer.TruncatedNormal(
-                            scale=0.02)),
-                    bias_attr=fluid.ParamAttr(
-                        name="cls_out_b_%d" % i,
-                        initializer=fluid.initializer.Constant(0.)),
-                    act="softmax"))
-
-        return probs
-
-    def _add_label(self):
-        label = fluid.layers.data(
-            name="label", shape=[self.num_classes], dtype='int64')
-        return [label]
-
-    def _add_loss(self):
-        label_split = fluid.layers.split(
-            self.labels[0], self.num_classes, dim=-1)
-        total_loss = fluid.layers.fill_constant(
-            shape=[1], value=0.0, dtype='float64')
-        for index, probs in enumerate(self.outputs):
-            ce_loss = fluid.layers.cross_entropy(
-                input=probs, label=label_split[index])
-            total_loss += fluid.layers.reduce_sum(ce_loss)
-        loss = fluid.layers.mean(x=total_loss)
-        return loss
-
-    def _add_metrics(self):
-        label_split = fluid.layers.split(
-            self.labels[0], self.num_classes, dim=-1)
-        # metrics change to auc of every class
-        eval_list = []
-        for index, probs in enumerate(self.outputs):
-            current_auc, _, _ = fluid.layers.auc(
-                input=probs, label=label_split[index])
-            eval_list.append(current_auc)
-        return eval_list
-
-    def _calculate_metrics(self, run_states):
-        loss_sum = acc_sum = run_examples = 0
-        run_step = run_time_used = 0
-        for run_state in run_states:
-            run_examples += run_state.run_examples
-            run_step += run_state.run_step
-            loss_sum += np.mean(
-                run_state.run_results[-1]) * run_state.run_examples
-        auc_list = run_states[-1].run_results[:-1]
-
-        run_time_used = time.time() - run_states[0].run_time_begin
-        avg_loss = loss_sum / (run_examples * self.num_classes)
-        run_speed = run_step / run_time_used
-
-        # The first key will be used as main metrics to update the best model
-        scores = OrderedDict()
-        for metric in self.metrics_choices:
-            if metric == "auc":
-                scores["auc"] = np.mean(auc_list)
-                # NOTE: for MultiLabelClassifierTask, the metrics will be used to evaluate all the label
-                #      and their mean value will also be reported.
-                for index, auc in enumerate(auc_list):
-                    scores["auc_" + self.class_name[index]] = auc_list[index][0]
-            else:
-                raise ValueError("Not Support Metric: \"%s\"" % metric)
-        return scores, avg_loss, run_speed
-
-    @property
-    def fetch_list(self):
-        if self.is_train_phase or self.is_test_phase:
-            return [metric.name for metric in self.metrics] + [self.loss.name]
-        return self.outputs
-
-
-class RegressionTask(BasicTask):
-    def __init__(self,
-                 feature,
-                 feed_list,
-                 data_reader,
-                 startup_program=None,
-                 config=None,
-                 hidden_units=None,
-                 metrics_choices="default"):
-        if metrics_choices == "default":
-            metrics_choices = ["spearman"]
-
-        main_program = feature.block.program
-        super(RegressionTask, self).__init__(
-            data_reader=data_reader,
-            main_program=main_program,
-            feed_list=feed_list,
-            startup_program=startup_program,
-            config=config,
-            metrics_choices=metrics_choices)
-        self.feature = feature
-        self.hidden_units = hidden_units
-
-    def _build_net(self):
-        cls_feats = fluid.layers.dropout(
-            x=self.feature,
-            dropout_prob=0.1,
-            dropout_implementation="upscale_in_train")
-
-        if self.hidden_units is not None:
-            for n_hidden in self.hidden_units:
-                cls_feats = fluid.layers.fc(
-                    input=cls_feats, size=n_hidden, act="relu")
-
-        logits = fluid.layers.fc(
-            input=cls_feats,
-            size=1,
-            param_attr=fluid.ParamAttr(
-                name="cls_out_w",
-                initializer=fluid.initializer.TruncatedNormal(scale=0.02)),
-            bias_attr=fluid.ParamAttr(
-                name="cls_out_b", initializer=fluid.initializer.Constant(0.)),
-            act=None)
-
-        return [logits]
-
-    def _add_label(self):
-        return [fluid.layers.data(name="label", dtype="float32", shape=[1])]
-
-    def _add_loss(self):
-        cost = fluid.layers.square_error_cost(
-            input=self.outputs[0], label=self.labels[0])
-        return fluid.layers.mean(x=cost)
-
-    def _add_metrics(self):
-        return []
-
-    @property
-    def fetch_list(self):
-        if self.is_train_phase or self.is_test_phase:
-            return [self.labels[0].name, self.outputs[0].name
-                    ] + [metric.name
-                         for metric in self.metrics] + [self.loss.name]
-        return [output.name for output in self.outputs]
-
-    def _calculate_metrics(self, run_states):
-        loss_sum = run_examples = 0
-        run_step = run_time_used = 0
-        all_labels = np.array([])
-        all_infers = np.array([])
-        for run_state in run_states:
-            run_examples += run_state.run_examples
-            run_step += run_state.run_step
-            loss_sum += np.mean(
-                run_state.run_results[-1]) * run_state.run_examples
-            np_labels = run_state.run_results[0]
-            np_infers = run_state.run_results[1]
-            all_labels = np.hstack((all_labels, np_labels.reshape([-1])))
-            all_infers = np.hstack((all_infers, np_infers.reshape([-1])))
-
-        run_time_used = time.time() - run_states[0].run_time_begin
-        avg_loss = loss_sum / run_examples
-        run_speed = run_step / run_time_used
-
-        # The first key will be used as main metrics to update the best model
-        scores = OrderedDict()
-
-        for metric in self.metrics_choices:
-            if metric == "spearman":
-                spearman_correlations = spearmanr(all_labels, all_infers)[0]
-                scores["spearman"] = spearman_correlations
-            else:
-                raise ValueError("Not Support Metric: \"%s\"" % metric)
-        return scores, avg_loss, run_speed
-
-
-class ReadingComprehensionTask(BasicTask):
-    def __init__(self,
-                 feature,
-                 feed_list,
-                 data_reader,
-                 startup_program=None,
-                 config=None,
-                 metrics_choices=None):
-
-        main_program = feature.block.program
-        super(ReadingComprehensionTask, self).__init__(
-            data_reader=data_reader,
-            main_program=main_program,
-            feed_list=feed_list,
-            startup_program=startup_program,
-            config=config,
-            metrics_choices=metrics_choices)
-        self.feature = feature
-
-    def _build_net(self):
-        if self.is_predict_phase:
-            self.unique_id = fluid.layers.data(
-                name="start_positions",
-                shape=[-1, 1],
-                lod_level=0,
-                dtype="int64")
-
-        logits = fluid.layers.fc(
-            input=self.feature,
-            size=2,
-            num_flatten_dims=2,
-            param_attr=fluid.ParamAttr(
-                name="cls_seq_label_out_w",
-                initializer=fluid.initializer.TruncatedNormal(scale=0.02)),
-            bias_attr=fluid.ParamAttr(
-                name="cls_seq_label_out_b",
-                initializer=fluid.initializer.Constant(0.)))
-
-        logits = fluid.layers.transpose(x=logits, perm=[2, 0, 1])
-        start_logits, end_logits = fluid.layers.unstack(x=logits, axis=0)
-
-        batch_ones = fluid.layers.fill_constant_batch_size_like(
-            input=start_logits, dtype='int64', shape=[1], value=1)
-        num_seqs = fluid.layers.reduce_sum(input=batch_ones)
-
-        return [start_logits, end_logits, num_seqs]
-
-    def _add_label(self):
-        start_positions = fluid.layers.data(
-            name="start_positions", shape=[-1, 1], lod_level=0, dtype="int64")
-        end_positions = fluid.layers.data(
-            name="end_positions", shape=[-1, 1], lod_level=0, dtype="int64")
-        return [start_positions, end_positions]
-
-    def _add_loss(self):
-        start_positions = self.labels[0]
-        end_positions = self.labels[1]
-
-        start_logits = self.outputs[0]
-        end_logits = self.outputs[1]
-
-        start_loss = fluid.layers.softmax_with_cross_entropy(
-            logits=start_logits, label=start_positions)
-        start_loss = fluid.layers.mean(x=start_loss)
-        end_loss = fluid.layers.softmax_with_cross_entropy(
-            logits=end_logits, label=end_positions)
-        end_loss = fluid.layers.mean(x=end_loss)
-        total_loss = (start_loss + end_loss) / 2.0
-        return total_loss
-
-    def _add_metrics(self):
-        return []
-
-    @property
-    def feed_list(self):
-        feed_list = [varname for varname in self._base_feed_list]
-        if self.is_train_phase:
-            feed_list += [self.labels[0].name, self.labels[1].name]
-        elif self.is_predict_phase:
-            feed_list += [self.unique_id.name]
-        return feed_list
-
-    @property
-    def fetch_list(self):
-        if self.is_train_phase:
-            return [metric.name for metric in self.metrics
-                    ] + [self.loss.name, self.outputs[-1].name]
-        elif self.is_predict_phase:
-            return [self.unique_id.name
-                    ] + [output.name for output in self.outputs]
-
-    def _calculate_metrics(self, run_states):
-        total_cost, total_num_seqs = [], []
-        run_step = run_time_used = run_examples = 0
-        for run_state in run_states:
-            np_loss = run_state.run_results[0]
-            np_num_seqs = run_state.run_results[1]
-            total_cost.extend(np_loss * np_num_seqs)
-            total_num_seqs.extend(np_num_seqs)
-            run_examples += run_state.run_examples
-            run_step += run_state.run_step
-
-        run_time_used = time.time() - run_states[0].run_time_begin
-        run_speed = run_step / run_time_used
-        avg_loss = np.sum(total_cost) / np.sum(total_num_seqs)
-
-        scores = OrderedDict()
-        # If none of metrics has been implemented, loss will be used to evaluate.
-        return scores, avg_loss, run_speed
